refactor(cli): split the learn candidate front end under the file-size limit
`scripts/cli/learn-command.js` grew from 105 to 788 lines in this branch,
past the 700-line hard limit in `.claude/rules/coding-standards.md`. The
grandfathering clause there is scoped to files already oversized under
`scripts/lib/`; this one was pushed past the limit by the queue
unification itself, so trimming it is cleaning up after that move.

Two flat siblings, both a verbatim move:

- `scripts/cli/learn-candidate-queue.js` — everything the candidate
  subcommands read from disk: this project's cards, the id lookup, and
  the newest materialization manifest's draft answers.
- `scripts/cli/learn-candidate-prose.js` — everything they say: the verb
  ↔ matrix-action mapping, the inbox ordering, the per-status prose and
  every refusal message. Pure over a card and an engine result.

`learn-command.js` keeps the entry and dispatch layer and stays the only
module `cli.js` requires; `runLearnCommand` remains its only export. No
behavior, flag, or JSON shape changes, so no spec or guide moves with it.
Every `../lib` import stays lazy inside its function body, so `learn
diary` still does not pay for the curator's module graph.

Flag reads deliberately all stay in `learn-command.js`:
`tests/node/test-cli-manifest.js` derives the `learn` live flag set by
scanning source, and a read that drifted into a sibling would shrink the
derived set while the ⊆ assertion still passed. That derivation now
concatenates all three files so the seam cannot open silently.

Addresses review thread on scripts/cli/learn-command.js:712.
